refactor(core): extract `Global::validate_pass_timestamp_writes`
